chore: refactor tool loop into multiple readable files
