+++ app/agent/tool_loop/context_budget.py
@@ -1,41 +1,12 @@
-"""Shared tool-calling ReAct primitives.
-
-This module owns the provider-agnostic machinery for running a "think → call
-tools → observe" loop against the registered tool set: parallel tool execution,
-provider-specific assistant/tool-result message shaping, and the context-window
-budget enforcement that keeps long loops under each model's prompt limit.
-
-Two consumers build on top of it:
-
-* :mod:`app.agent.nodes.investigate` — the investigation agent layers evidence
-  collection, seed calls, and diagnosis parsing on top of these helpers and its
-  own loop orchestration.
-* the interactive shell's tool-gathering pass — uses :func:`run_tool_calling_loop`
-  to let the REPL assistant pull live data from the *same* registered tools the
-  investigation uses before composing a conversational answer.
-
-Keeping the loop primitives here (rather than private to ``investigation.py``)
-means both surfaces share one implementation of the subtle, well-tested context
-budgeting and provider message shaping.
-"""
+"""Context-window budgeting for shared agent tool loops."""
 
 from __future__ import annotations
 
 import json
 import logging
-from collections.abc import Callable
-from concurrent.futures import Future, ThreadPoolExecutor, as_completed
-from dataclasses import dataclass, field
 from typing import Any
 
-from app.services.agent_llm_client import ToolCall
-from app.tools.registered_tool import RegisteredTool
-from app.utils.tool_trace import redact_sensitive
-
 logger = logging.getLogger(__name__)
-
-_TOOL_EXECUTOR_WORKERS = 10
-_UNSET: object = object()  # sentinel distinguishing "not yet started" from a None tool result
 
 # Defensive context-window ceiling. Below this we never trim; above this we
 # drop the oldest tool_use/tool_result pair until back under the ceiling.
@@ -95,11 +66,6 @@ _TRUNCATION_SAFETY_TOKENS = 2_000
 # already consume the whole ceiling, we still leave at least this much so the
 # truncated message carries some signal instead of being blanked.
 _TRUNCATION_MIN_TOKENS = 1_000
-
-
-# Callback type: called with (event_kind, data_dict) during the agent loop.
-# event_kind values: "tool_start", "tool_end", "llm_start", "agent_start", "agent_end"
-AgentEventCallback = Callable[[str, dict[str, Any]], None]
 
 
 def _context_budget_ceiling_for_model(model: str | None) -> int:
@@ -360,255 +326,3 @@ def _enforce_context_budget(
         logger.warning(
             "[agent] trimmed oldest tool pair to fit context budget (ceiling=%d)", ceiling
         )
-
-
-def _build_synthetic_assistant_tool_call_msg(
-    llm: Any,
-    tool_calls: list[ToolCall],
-) -> dict[str, Any]:
-    """Build an assistant message that looks like the LLM requested these tool calls.
-
-    This lets us inject pre-seeded tool results into the conversation in a format
-    the LLM client already understands, without adding special-case handling.
-    """
-    from app.services.agent_llm_client import (
-        AnthropicAgentClient,
-        BedrockConverseAgentClient,
-        CLIBackedAgentClient,
-        OpenAIAgentClient,
-    )
-
-    if isinstance(llm, BedrockConverseAgentClient):
-        from app.services.bedrock_converse import build_assistant_tool_use_message
-
-        return build_assistant_tool_use_message(tool_calls)
-
-    if isinstance(llm, AnthropicAgentClient):
-        content = [
-            {
-                "type": "tool_use",
-                "id": tc.id,
-                "name": tc.name,
-                "input": tc.input,
-            }
-            for tc in tool_calls
-        ]
-        return {"role": "assistant", "content": content}
-
-    if isinstance(llm, OpenAIAgentClient):
-        return {
-            "role": "assistant",
-            "content": None,
-            "tool_calls": [
-                {
-                    "id": tc.id,
-                    "type": "function",
-                    "function": {"name": tc.name, "arguments": json.dumps(tc.input)},
-                }
-                for tc in tool_calls
-            ],
-        }
-
-    if isinstance(llm, CLIBackedAgentClient):
-        return llm.build_assistant_message("", tool_calls)
-
-    # Fallback: plain text summary
-    names = ", ".join(tc.name for tc in tool_calls)
-    return {"role": "assistant", "content": f"I will start by querying: {names}"}
-
-
-def _run_parallel(
-    tool_calls: list[ToolCall],
-    tools: list[RegisteredTool],
-    resolved_integrations: dict[str, Any],
-) -> list[Any]:
-    tool_map = {t.name: t for t in tools}
-
-    def _call(tc: ToolCall) -> Any:
-        tool = tool_map.get(tc.name)
-        if tool is None:
-            return {"error": f"unknown tool: {tc.name}"}
-        try:
-            validation_error = tool.validate_public_input(tc.input)
-            if validation_error:
-                return {"error": validation_error}
-            injected = tool.extract_params(resolved_integrations)
-            kwargs = {**injected, **tc.input}
-            return tool.run(**kwargs)
-        except Exception as exc:
-            logger.warning("[tool:%s] failed: %s", tc.name, exc)
-            return {"error": str(exc)}
-
-    if len(tool_calls) == 1:
-        return [_call(tool_calls[0])]
-
-    results: list[Any] = [_UNSET] * len(tool_calls)
-    submitted: dict[
-        Future[Any], int
-    ] = {}  # future -> index, built incrementally to survive partial submit
-    try:
-        with ThreadPoolExecutor(max_workers=min(_TOOL_EXECUTOR_WORKERS, len(tool_calls))) as pool:
-            for i, tc in enumerate(tool_calls):
-                submitted[pool.submit(_call, tc)] = i
-            for fut in as_completed(submitted):
-                try:
-                    results[submitted[fut]] = fut.result()
-                except Exception as fut_exc:  # noqa: BLE001  # lgtm[py/catch-base-exception]
-                    results[submitted[fut]] = {"error": str(fut_exc)}
-    except RuntimeError as exc:
-        # interpreter is shutting down; executor.__exit__ has already waited for submitted futures
-        logger.warning("[_run_parallel] RuntimeError – falling back to sequential: %s", exc)
-        for fut, i in submitted.items():
-            if results[i] is _UNSET and fut.done():
-                try:
-                    results[i] = fut.result()
-                except Exception as fut_exc:  # noqa: BLE001  # lgtm[py/catch-base-exception]
-                    results[i] = {"error": str(fut_exc)}
-        for i, tc in enumerate(tool_calls):
-            if results[i] is _UNSET:
-                results[i] = _call(tc)
-    return results
-
-
-def _public_tool_input(value: dict[str, Any]) -> dict[str, Any]:
-    redacted = redact_sensitive(value)
-    return {
-        key: item
-        for key, item in redacted.items()
-        if item != "[runtime object]" and item != "[redacted]"
-    }
-
-
-def _tool_source(tools: list[RegisteredTool], tool_name: str) -> str:
-    for tool in tools:
-        if tool.name == tool_name:
-            return str(tool.source)
-    return "unknown"
-
-
-def _summarise(output: Any) -> str:
-    if isinstance(output, dict) and "error" in output:
-        return f"error: {output['error']}"
-    text = json.dumps(output, default=str)
-    return text[:120] + "…" if len(text) > 120 else text
-
-
-def _build_assistant_msg(llm: Any, response: Any) -> dict[str, Any]:
-    from app.services.agent_llm_client import AnthropicAgentClient, BedrockConverseAgentClient
-
-    if isinstance(llm, (AnthropicAgentClient, BedrockConverseAgentClient)):
-        return llm.build_assistant_message(response.raw_content)
-    # Use raw_content when set — preserves provider-specific fields such as
-    # Gemini's thought_signature that must be echoed back in the next request.
-    if response.raw_content is not None:
-        return response.raw_content  # type: ignore[no-any-return]
-    result: dict[str, Any] = llm.build_assistant_message(response.content, response.tool_calls)
-    return result
-
-
-def _build_tool_result_messages(
-    llm: Any,
-    tool_calls: list[ToolCall],
-    results: list[Any],
-) -> list[dict[str, Any]]:
-    from app.services.agent_llm_client import AnthropicAgentClient, OpenAIAgentClient
-
-    if isinstance(llm, AnthropicAgentClient):
-        return [llm.build_tool_result_message(tool_calls, results)]
-    if isinstance(llm, OpenAIAgentClient):
-        return llm.build_tool_result_messages(tool_calls, results)
-    return [llm.build_tool_result_message(tool_calls, results)]
-
-
-@dataclass
-class ToolLoopResult:
-    """Outcome of :func:`run_tool_calling_loop`.
-
-    ``messages`` is the full conversation (mutated in place and returned for
-    convenience), ``final_text`` is the assistant's last no-tool-call turn (the
-    conversational answer, empty when the loop hit the iteration cap), and
-    ``executed`` is the ordered list of ``(tool_call, output)`` pairs run during
-    the loop.
-    """
-
-    messages: list[dict[str, Any]]
-    final_text: str
-    executed: list[tuple[ToolCall, Any]] = field(default_factory=list)
-    hit_iteration_cap: bool = False
-
-
-def run_tool_calling_loop(
-    *,
-    llm: Any,
-    system: str,
-    messages: list[dict[str, Any]],
-    tools: list[RegisteredTool],
-    resolved_integrations: dict[str, Any],
-    max_iterations: int,
-    on_event: AgentEventCallback | None = None,
-) -> ToolLoopResult:
-    """Run a generic think → call-tools → observe loop and return its outcome.
-
-    Unlike :class:`app.agent.nodes.investigate.ConnectedInvestigationAgent`, this is
-    a plain conversational loop: it does not seed tool calls, collect evidence,
-    or parse a diagnosis. It exists so non-investigation surfaces (currently the
-    interactive shell's tool-gathering pass) can call the *same* registered tools
-    the investigation uses, with the same provider message shaping and context
-    budgeting.
-
-    ``on_event`` mirrors the investigation agent's callback contract so callers
-    can render ``tool_start`` / ``tool_end`` activity live.
-    """
-
-    def _emit(kind: str, data: dict[str, Any]) -> None:
-        if on_event is not None:
-            try:
-                on_event(kind, data)
-            except Exception:  # noqa: BLE001 — event rendering must never break the loop
-                logger.debug("[tool_loop] on_event(%s) raised; ignoring", kind, exc_info=True)
-
-    tool_schemas = llm.tool_schemas(tools)
-    ceiling = _context_budget_ceiling_for_model(getattr(llm, "_model", None))
-    executed: list[tuple[ToolCall, Any]] = []
-    final_text = ""
-    hit_cap = True
-
-    for iteration in range(max_iterations):
-        _emit("llm_start", {"iteration": iteration})
-        _enforce_context_budget(messages, system=system, tools=tool_schemas, ceiling=ceiling)
-        response = llm.invoke(messages, system=system, tools=tool_schemas)
-        messages.append(_build_assistant_msg(llm, response))
-
-        if not response.has_tool_calls:
-            final_text = response.content or ""
-            hit_cap = False
-            break
-
-        for tc in response.tool_calls:
-            _emit(
-                "tool_start", {"id": tc.id, "name": tc.name, "input": _public_tool_input(tc.input)}
-            )
-
-        results = _run_parallel(response.tool_calls, tools, resolved_integrations)
-        messages.extend(_build_tool_result_messages(llm, response.tool_calls, results))
-
-        for tc, output in zip(response.tool_calls, results):
-            executed.append((tc, output))
-            _emit(
-                "tool_end",
-                {"id": tc.id, "name": tc.name, "output": redact_sensitive(output)},
-            )
-
-    return ToolLoopResult(
-        messages=messages,
-        final_text=final_text,
-        executed=executed,
-        hit_iteration_cap=hit_cap,
-    )
-
-
-__all__ = [
-    "AgentEventCallback",
-    "ToolLoopResult",
-    "run_tool_calling_loop",
-]
